refactor: SwiftDataモデルクラスを個別ファイルに分割
コードの可読性とメンテナンス性向上のため、SwiftDataモデルクラスを個別ファイルに分割しました。

- SwiftDataBook.swift: 絵本モデルを独立ファイルに分割
- SwiftDataUser.swift: 利用者モデルを独立ファイルに分割
- SwiftDataLoan.swift: 貸出モデルを独立ファイルに分割
- SwiftDataClassGroup.swift: クラス（組）モデルを独立ファイルに分割

各モデルクラスには詳細なドキュメントコメントを追加し、
SwiftDataの制約や設計判断についても明記しました。
