+++ PictureBookLendingAdminApp/PictureBookLendingInfrastructure/Sources/PictureBookLendingInfrastructure/Gateways/GoogleBooksAPIClient.swift
@@ -157,41 +157,3 @@ public struct GoogleBooksAPIClient: Sendable {
         )
     }
 }
-
-// MARK: - Google Books API Response Models
-
-/// Google Books API レスポンスルート
-struct VolumesResponse: Decodable {
-    let items: [Volume]?
-}
-
-/// 書籍ボリューム
-struct Volume: Decodable {
-    let volumeInfo: VolumeInfo
-}
-
-/// 書籍詳細情報
-struct VolumeInfo: Decodable {
-    let title: String?
-    let authors: [String]?
-    let publisher: String?
-    let publishedDate: String?
-    let description: String?
-    let pageCount: Int?
-    let categories: [String]?
-    let imageLinks: ImageLinks?
-    let industryIdentifiers: [IndustryIdentifier]?
-    let infoLink: URL?
-}
-
-/// 画像リンク
-struct ImageLinks: Decodable {
-    let smallThumbnail: String?
-    let thumbnail: String?
-}
-
-/// 業界識別子（ISBN等）
-struct IndustryIdentifier: Decodable {
-    let type: String
-    let identifier: String
-}
