+++ PictureBookLendingAdminApp/PictureBookLendingInfrastructure/Sources/PictureBookLendingInfrastructure/Repositories/SwiftDataBookRepository.swift
@@ -172,47 +172,3 @@ public class SwiftDataBookRepository: BookRepositoryProtocol {
         }
     }
 }
-
-/// SwiftData用の絵本モデル
-///
-/// SwiftDataで永続化するための絵本モデル
-@Model
-final public class SwiftDataBook {
-    public var id: UUID
-    public var title: String
-    public var author: String
-    public var isbn13: String?
-    public var publisher: String?
-    public var publishedDate: String?
-    public var bookDescription: String?  // SwiftDataでは'description'は予約語なので別名
-    public var thumbnailURLString: String?  // URLは直接保存できないので文字列で保存
-    public var targetAge: Int?
-    public var pageCount: Int?
-    public var categories: [String]
-    
-    public init(
-        id: UUID,
-        title: String,
-        author: String,
-        isbn13: String? = nil,
-        publisher: String? = nil,
-        publishedDate: String? = nil,
-        bookDescription: String? = nil,
-        thumbnailURLString: String? = nil,
-        targetAge: Int? = nil,
-        pageCount: Int? = nil,
-        categories: [String] = []
-    ) {
-        self.id = id
-        self.title = title
-        self.author = author
-        self.isbn13 = isbn13
-        self.publisher = publisher
-        self.publishedDate = publishedDate
-        self.bookDescription = bookDescription
-        self.thumbnailURLString = thumbnailURLString
-        self.targetAge = targetAge
-        self.pageCount = pageCount
-        self.categories = categories
-    }
-}
